+++ cookbooks/chub_logstash/attributes/default.rb
@@ -1,0 +1,3 @@
+default[:chub_logstash][:user]		"logstash"
+default[:chub_logstash][:base_dir]	"/opt/logstash"
+default[:chub_logstash][:cert_dir]	"#{node[:chub_logstash][:base_dir]}/ssl"